commit by xuhuangchao

+++ malldemo/oomall/discountModule/src/main/java/xmu/oomall/discount/controller/PresaleController.java
@@ -10,6 +10,7 @@ import xmu.oomall.discount.service.Impl.PresaleServiceImpl;
 import xmu.oomall.util.ResponseUtil;
 
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
@@ -18,6 +19,11 @@ public class PresaleController {
     @Autowired
     private PresaleServiceImpl presaleService;
 
+    /**
+     * 判断presaleRule是否符合规范
+     * @param presaleRule
+     * @return
+     */
     private Object validate(PresaleRule presaleRule) {
         Integer id = presaleRule.getGoodsId();
         if (StringUtils.isEmpty(String.valueOf(id))) {
@@ -59,12 +65,36 @@ public class PresaleController {
      * @Date: 2019/12/10
      */
     @PutMapping("/presaleRule/{id}")
-    public Object update(PresaleRule presaleRule){
+    public Object update(@PathVariable Integer id,PresaleRule presaleRule){
+        Object error = validate(presaleRule);
+        if (error != null) {
+            return error;
+        }
+        presaleRule.setId(id);
+        PresaleRule ruleInDB=presaleService.findById(id);
+        Boolean statusCode=ruleInDB.getStatusCode();
+        LocalDateTime beginTime=ruleInDB.getStartTime();
+        LocalDateTime endTime=ruleInDB.getEndTime();
+        LocalDateTime nowTime=LocalDateTime.now();
 
-        if (presaleService.update(presaleRule)== 1) {
+        //预售规则未失效
+        if(statusCode==true) {
+            if ((nowTime.compareTo(beginTime) >= 0) && (nowTime.compareTo(endTime) <= 0)) {
+                //在预售开始到结束时间内不能改动信息
+                return ResponseUtil.updatedDataFailed();
+            } else {
+                //预售未开始或者已经结束可以修改信息
+                if (presaleService.update(presaleRule) == 0) {
+                    return ResponseUtil.updatedDataFailed();
+                }
+                return ResponseUtil.ok(presaleRule);
+            }
+        }else{//预售规则已经失效
+            if (presaleService.update(presaleRule) == 0) {
+                return ResponseUtil.updatedDataFailed();
+            }
             return ResponseUtil.ok(presaleRule);
         }
-        return ResponseUtil.updatedDataFailed();
     }
 
 
@@ -77,24 +107,25 @@ public class PresaleController {
      */
     @DeleteMapping("/presaleRule/{id}")
     public Object delete(@PathVariable Integer id){
-        presaleService.delete(id);
-        return ResponseUtil.ok();
+        PresaleRule ruleInDB=presaleService.findById(id);
+        Boolean statusCode=ruleInDB.getStatusCode();
+        LocalDateTime beginTime=ruleInDB.getStartTime();
+        LocalDateTime endTime=ruleInDB.getEndTime();
+        LocalDateTime nowTime=LocalDateTime.now();
+        //预售规则未失效
+        if(statusCode==true) {
+            //在预售开始到结束时间内不能删除优惠券
+            if ((nowTime.compareTo(beginTime) >= 0) && (nowTime.compareTo(endTime) <= 0)) {
+                return ResponseUtil.fail();
+            }
+            presaleService.delete(id);
+            return ResponseUtil.ok();
+        }else {//预售规则已失效
+            presaleService.delete(id);
+            return ResponseUtil.ok();
+        }
     }
 
-
-    /**
-     * Order调discount获得预售订单的定金和尾款
-     * @param order
-     * @param maxPayTime
-     * @return
-     */
-
-    @GetMapping("/calcWithPrePayment")
-    public Object getDepositAndFinalPay(Order order,Integer maxPayTime){
-
-        List<BigDecimal> priceList=presaleService.getDepositAndFinalPay(order,maxPayTime);
-        return ResponseUtil.ok(priceList);
-    }
 
 
 }
